Rename file to match function
