+++ packages/graph-explorer/src/connector/gremlin/useGremlin.ts
@@ -5,7 +5,7 @@ import fetchNeighborsCount from "./queries/fetchNeighborsCount";
 import fetchSchema from "./queries/fetchSchema";
 import fetchVertexTypeCounts from "./queries/fetchVertexTypeCounts";
 import keywordSearch from "./queries/keywordSearch";
-import { fetchDatabaseRequest } from "../useGEFetch";
+import { fetchDatabaseRequest } from "../fetchDatabaseRequest";
 import { GraphSummary } from "./types";
 import { v4 } from "uuid";
